test: organize model tests
