+++ test/unit/models/qwen.test.mjs
@@ -1,19 +1,11 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import { BaseEmbeddingModel } from "../../dist/engine/models/embeddings.js";
-import {
-  getEmbeddingModelCatalogEntry,
-  listEmbeddingModels,
-} from "../../dist/engine/models/catalog.js";
-import { createEmbeddingModel } from "../../dist/engine/models/factory.js";
-import { LlamaCppEmbeddingModel } from "../../dist/engine/models/backends/llama-cpp.js";
-import { Model2VecEmbeddingModel } from "../../dist/engine/models/backends/model2vec.js";
+import { getEmbeddingModelCatalogEntry } from "../../../dist/engine/models/catalog.js";
 import {
   Qwen37TextEmbeddingModel,
   Qwen3VlEmbeddingModel,
   QwenTextEmbeddingV4Model,
-} from "../../dist/engine/models/backends/qwen.js";
-import { TransformersJsEmbeddingModel } from "../../dist/engine/models/backends/transformers-js.js";
+} from "../../../dist/engine/models/backends/qwen.js";
 
 const vector = (dimension, value = 0.25) => Array(dimension).fill(value);
 const qwenTextEntry = getEmbeddingModelCatalogEntry("qwen/text-embedding-v4");
@@ -38,105 +30,6 @@ function jsonResponse(body, init = {}) {
     headers: { "content-type": "application/json", ...init.headers },
   });
 }
-
-class StubEmbeddingModel extends BaseEmbeddingModel {
-  info = {
-    reference: "test/stub",
-    provider: "test",
-    name: "stub",
-    dimension: 2,
-    metric: "cosine",
-    inputKinds: ["text", "image"],
-    limits: { maxBatchSize: 2, maxImageBytes: 3 },
-  };
-  result = [[1, 0]];
-  seenPurpose;
-  seenSignal;
-
-  async doEmbed(_contents, options) {
-    this.seenPurpose = options.purpose;
-    this.seenSignal = options.signal;
-    return { vectors: this.result, truncated: [] };
-  }
-}
-
-test("embedding base class validates inputs, provider outputs, and purpose defaults", async () => {
-  const model = new StubEmbeddingModel();
-  assert.deepEqual(await model.embed([{ kind: "text", text: "hello" }]), {
-    vectors: [[1, 0]],
-    truncated: [],
-  });
-  assert.equal(model.seenPurpose, "document");
-  const signal = new AbortController().signal;
-  await model.embed([{ kind: "text", text: "query" }], {
-    purpose: "query",
-    signal,
-  });
-  assert.equal(model.seenPurpose, "query");
-  assert.equal(model.seenSignal, signal);
-  await model.dispose();
-
-  await assert.rejects(model.embed([]), /at least one/);
-  await assert.rejects(
-    model.embed([
-      { kind: "text", text: "one" },
-      { kind: "text", text: "two" },
-      { kind: "text", text: "three" },
-    ]),
-    /batch size/,
-  );
-  await assert.rejects(
-    model.embed([{ kind: "text", text: "  " }]),
-    /must not be empty/,
-  );
-  await assert.rejects(
-    model.embed([{ kind: "image", data: new Uint8Array(), format: "png" }]),
-    /must not be empty/,
-  );
-  await assert.rejects(
-    model.embed([{ kind: "image", data: new Uint8Array([1]), format: "" }]),
-    /include a format/,
-  );
-  await assert.rejects(
-    model.embed([
-      { kind: "image", data: new Uint8Array([1, 2, 3, 4]), format: "png" },
-    ]),
-    /exceeds model limit/,
-  );
-
-  model.info.inputKinds = ["text"];
-  await assert.rejects(
-    model.embed([{ kind: "image", data: new Uint8Array([1]), format: "png" }]),
-    /does not support/,
-  );
-  model.info.inputKinds = ["text", "image"];
-
-  model.result = null;
-  await assert.rejects(
-    model.embed([{ kind: "text", text: "value" }]),
-    /non-array response/,
-  );
-  model.result = [];
-  await assert.rejects(
-    model.embed([{ kind: "text", text: "value" }]),
-    /wrong number/,
-  );
-  model.result = [null];
-  await assert.rejects(
-    model.embed([{ kind: "text", text: "value" }]),
-    /non-array vector/,
-  );
-  model.result = [[1]];
-  await assert.rejects(
-    model.embed([{ kind: "text", text: "value" }]),
-    /wrong dimension/,
-  );
-  model.result = [[1, Number.NaN]];
-  await assert.rejects(
-    model.embed([{ kind: "text", text: "value" }]),
-    /non-finite/,
-  );
-});
 
 test("Qwen text model sends ordered batches and validates all response shapes", async () => {
   assert.throws(
@@ -424,59 +317,4 @@ test("Qwen VL model validates images, encodes bytes, and accepts provider index 
         assert.rejects(model.embed([{ kind: "text", text: "value" }]), message),
     );
   }
-});
-
-test("embedding factory resolves catalog entries and rejects unknown models", () => {
-  const options = { apiKey: "secret", endpoint: "https://example.test" };
-  const expectedModelClasses = new Map([
-    ["local/embeddinggemma-300m", LlamaCppEmbeddingModel],
-    ["local/qwen3-embedding-0.6b", LlamaCppEmbeddingModel],
-    ["qwen/text-embedding-v4", QwenTextEmbeddingV4Model],
-    ["qwen/qwen3.7-text-embedding", Qwen37TextEmbeddingModel],
-    ["qwen/qwen3-vl-embedding", Qwen3VlEmbeddingModel],
-    ["local/bge-small-en-v1.5", TransformersJsEmbeddingModel],
-    ["local/all-minilm-l6-v2", TransformersJsEmbeddingModel],
-    ["local/potion-base-8m", Model2VecEmbeddingModel],
-    ["local/potion-code-16m-v2", Model2VecEmbeddingModel],
-    ["local/multilingual-e5-small", TransformersJsEmbeddingModel],
-    ["local/jina-embeddings-v2-base-code", TransformersJsEmbeddingModel],
-    ["local/gte-modernbert-base", TransformersJsEmbeddingModel],
-    ["local/nomic-embed-text-v1.5", TransformersJsEmbeddingModel],
-  ]);
-  const catalogEntries = listEmbeddingModels();
-
-  assert.deepEqual(
-    catalogEntries.map((entry) => entry.reference).sort(),
-    [...expectedModelClasses.keys()].sort(),
-  );
-
-  for (const entry of catalogEntries) {
-    const model = createEmbeddingModel(entry.reference, options);
-    const ExpectedModel = expectedModelClasses.get(entry.reference);
-
-    assert.ok(
-      model instanceof ExpectedModel,
-      `${entry.reference} resolved to ${model.constructor.name}`,
-    );
-    assert.equal(model.info.reference, entry.reference);
-    assert.equal(model.info.dimension, entry.dimension);
-    assert.equal(model.info.metric, entry.metric);
-  }
-
-  assert.throws(
-    () => createEmbeddingModel("missing", options),
-    /not in the zvec-grep catalog/,
-  );
-  assert.throws(
-    () => createEmbeddingModel("invalid", options),
-    /not in the zvec-grep catalog/,
-  );
-  assert.throws(
-    () => createEmbeddingModel("unknown/missing", options),
-    /not in the zvec-grep catalog/,
-  );
-  assert.throws(
-    () => createEmbeddingModel("local/missing", options),
-    /not in the zvec-grep catalog/,
-  );
 });
